dljobs/CNN/requirements.txt
pynvml==11.0.0
PyYAML==5.4.1
tensorflow_datasets==4.4.0